feat: US-001 - Extract the import preview into a dedicated dashboard-owned dialog surface
